Split up graphql resolvers into different files
